made some cleanup in Repository Project
