+++ CodeMaze-V2/internal_logs/internallog.txt
@@ -1463,3 +1463,79 @@
 2023-12-27 19:33:28.1870 Trace FileTarget(Name=logfile) has 2 layouts
 2023-12-27 19:33:28.1870 Trace FileTarget(Name=logfile): Using appenderFactory: NLog.Internal.FileAppenders.SingleProcessFileAppender+Factory
 2023-12-27 19:33:28.1870 Info Configuration initialized.
+2023-12-27 19:41:40.9246 Debug Message Template Auto Format enabled
+2023-12-27 19:41:40.9894 Trace FindReachableObject<NLog.Internal.IRenderable>:
+2023-12-27 19:41:40.9894 Trace Scanning ShortDateLayoutRenderer 'Layout Renderer: ${shortdate}'
+2023-12-27 19:41:40.9894 Debug FileTarget FilePathLayout with FilePathKind.Relative using AppDomain.BaseDirectory: D:\dotnet\dotNetApiCodeMaze\CodeMaze-V2\CodeMaze-V2\bin\Debug\net6.0\
+2023-12-27 19:41:40.9894 Trace FindReachableObject<NLog.Internal.IRenderable>:
+2023-12-27 19:41:40.9894 Trace Scanning LongDateLayoutRenderer 'Layout Renderer: ${longdate}'
+2023-12-27 19:41:40.9894 Trace FindReachableObject<NLog.Internal.IRenderable>:
+2023-12-27 19:41:40.9894 Trace Scanning LevelLayoutRenderer 'Layout Renderer: ${level}'
+2023-12-27 19:41:40.9894 Trace FindReachableObject<NLog.Internal.IRenderable>:
+2023-12-27 19:41:40.9894 Trace Scanning MessageLayoutRenderer 'Layout Renderer: ${message}'
+2023-12-27 19:41:41.0000 Debug Adding target NLog.Targets.FileTarget(Name=logfile)
+2023-12-27 19:41:41.0000 Info Registered target NLog.Targets.FileTarget(Name=logfile)
+2023-12-27 19:41:41.0000 Trace ParseRulesElement
+2023-12-27 19:41:41.0155 Info NLog, Version=5.0.0.0, Culture=neutral, PublicKeyToken=5120e14c03d0593c. File version: 5.2.5.2160. Product version: 5.2.5+7c014325443d65541200b698a50b9dd4ec30c7f0. GlobalAssemblyCache: False
+2023-12-27 19:41:41.0155 Debug Watching file-filter 'nlog.config' in directory: D:\dotnet\dotNetApiCodeMaze\CodeMaze-V2\CodeMaze-V2
+2023-12-27 19:41:41.0155 Debug --- NLog configuration dump ---
+2023-12-27 19:41:41.0155 Debug Targets:
+2023-12-27 19:41:41.0155 Debug FileTarget(Name=logfile)
+2023-12-27 19:41:41.0155 Debug Rules:
+2023-12-27 19:41:41.0155 Debug logNamePattern: (:All) levels: [ Debug Info Warn Error Fatal ] writeTo: [ logfile ]
+2023-12-27 19:41:41.0155 Debug --- End of NLog configuration dump ---
+2023-12-27 19:41:41.0155 Trace FindReachableObject<System.Object>:
+2023-12-27 19:41:41.0155 Trace Scanning LoggingRule 'logNamePattern: (:All) levels: [ Debug Info Warn Error Fatal ] writeTo: [ logfile ]'
+2023-12-27 19:41:41.0155 Trace  Scanning Property Targets 'System.Collections.Generic.List`1[NLog.Targets.Target]' System.Collections.Generic.IList`1[NLog.Targets.Target]
+2023-12-27 19:41:41.0155 Trace  Scanning FileTarget 'FileTarget(Name=logfile)'
+2023-12-27 19:41:41.0155 Trace   Scanning Property FileName '.\logs\${shortdate}_logfile.txt' NLog.Layouts.Layout
+2023-12-27 19:41:41.0155 Trace   Scanning SimpleLayout '.\logs\${shortdate}_logfile.txt'
+2023-12-27 19:41:41.0155 Trace    Scanning Property LayoutRenderers 'NLog.LayoutRenderers.LayoutRenderer[]' System.Collections.Generic.IEnumerable`1[NLog.LayoutRenderers.LayoutRenderer]
+2023-12-27 19:41:41.0302 Trace    Scanning LiteralLayoutRenderer 'Layout Renderer: ${literal}'
+2023-12-27 19:41:41.0302 Trace    Scanning ShortDateLayoutRenderer 'Layout Renderer: ${shortdate}'
+2023-12-27 19:41:41.0302 Trace    Scanning LiteralLayoutRenderer 'Layout Renderer: ${literal}'
+2023-12-27 19:41:41.0302 Trace   Scanning Property Layout '${longdate} ${level:uppercase=true} ${message}' NLog.Layouts.Layout
+2023-12-27 19:41:41.0302 Trace   Scanning SimpleLayout '${longdate} ${level:uppercase=true} ${message}'
+2023-12-27 19:41:41.0302 Trace    Scanning Property LayoutRenderers 'NLog.LayoutRenderers.LayoutRenderer[]' System.Collections.Generic.IEnumerable`1[NLog.LayoutRenderers.LayoutRenderer]
+2023-12-27 19:41:41.0302 Trace    Scanning LongDateLayoutRenderer 'Layout Renderer: ${longdate}'
+2023-12-27 19:41:41.0302 Trace    Scanning LiteralLayoutRenderer 'Layout Renderer: ${literal}'
+2023-12-27 19:41:41.0302 Trace    Scanning LevelLayoutRenderer 'Layout Renderer: ${level}'
+2023-12-27 19:41:41.0302 Trace    Scanning LiteralLayoutRenderer 'Layout Renderer: ${literal}'
+2023-12-27 19:41:41.0302 Trace    Scanning MessageLayoutRenderer 'Layout Renderer: ${message}'
+2023-12-27 19:41:41.0302 Trace  Scanning Property ChildRules 'System.Collections.Generic.List`1[NLog.Config.LoggingRule]' System.Collections.Generic.IList`1[NLog.Config.LoggingRule]
+2023-12-27 19:41:41.0302 Trace  Scanning Property Filters 'System.Collections.Generic.List`1[NLog.Filters.Filter]' System.Collections.Generic.IList`1[NLog.Filters.Filter]
+2023-12-27 19:41:41.0302 Info Validating config: TargetNames=logfile, ConfigItems=12, FilePath=D:\dotnet\dotNetApiCodeMaze\CodeMaze-V2\CodeMaze-V2\nlog.config
+2023-12-27 19:41:41.0302 Debug Unused target checking is started... Rule Count: 1, Target Count: 1
+2023-12-27 19:41:41.0302 Debug Unused target checking is completed. Total Rule Count: 1, Total Target Count: 1, Unused Target Count: 0
+2023-12-27 19:41:41.0302 Trace Initializing Layout Renderer: ${message}
+2023-12-27 19:41:41.0302 Trace Initializing Layout Renderer: ${literal}
+2023-12-27 19:41:41.0302 Trace Initializing Layout Renderer: ${level}
+2023-12-27 19:41:41.0302 Trace Initializing Layout Renderer: ${literal}
+2023-12-27 19:41:41.0302 Trace Initializing Layout Renderer: ${longdate}
+2023-12-27 19:41:41.0302 Trace Initializing ${longdate} ${level:uppercase=true} ${message}
+2023-12-27 19:41:41.0302 Trace FindReachableObject<NLog.Internal.IRenderable>:
+2023-12-27 19:41:41.0302 Trace Scanning SimpleLayout '${longdate} ${level:uppercase=true} ${message}'
+2023-12-27 19:41:41.0302 Trace  Scanning Property LayoutRenderers 'NLog.LayoutRenderers.LayoutRenderer[]' System.Collections.Generic.IEnumerable`1[NLog.LayoutRenderers.LayoutRenderer]
+2023-12-27 19:41:41.0302 Trace  Scanning LongDateLayoutRenderer 'Layout Renderer: ${longdate}'
+2023-12-27 19:41:41.0302 Trace  Scanning LiteralLayoutRenderer 'Layout Renderer: ${literal}'
+2023-12-27 19:41:41.0302 Trace  Scanning LevelLayoutRenderer 'Layout Renderer: ${level}'
+2023-12-27 19:41:41.0302 Trace  Scanning LiteralLayoutRenderer 'Layout Renderer: ${literal}'
+2023-12-27 19:41:41.0302 Trace  Scanning MessageLayoutRenderer 'Layout Renderer: ${message}'
+2023-12-27 19:41:41.0302 Trace Initializing Layout Renderer: ${literal}
+2023-12-27 19:41:41.0466 Trace Initializing Layout Renderer: ${shortdate}
+2023-12-27 19:41:41.0466 Trace Initializing Layout Renderer: ${literal}
+2023-12-27 19:41:41.0466 Trace Initializing .\logs\${shortdate}_logfile.txt
+2023-12-27 19:41:41.0466 Trace FindReachableObject<NLog.Internal.IRenderable>:
+2023-12-27 19:41:41.0466 Trace Scanning SimpleLayout '.\logs\${shortdate}_logfile.txt'
+2023-12-27 19:41:41.0466 Trace  Scanning Property LayoutRenderers 'NLog.LayoutRenderers.LayoutRenderer[]' System.Collections.Generic.IEnumerable`1[NLog.LayoutRenderers.LayoutRenderer]
+2023-12-27 19:41:41.0466 Trace  Scanning LiteralLayoutRenderer 'Layout Renderer: ${literal}'
+2023-12-27 19:41:41.0466 Trace  Scanning ShortDateLayoutRenderer 'Layout Renderer: ${shortdate}'
+2023-12-27 19:41:41.0466 Trace  Scanning LiteralLayoutRenderer 'Layout Renderer: ${literal}'
+2023-12-27 19:41:41.0466 Trace Initializing FileTarget(Name=logfile)
+2023-12-27 19:41:41.0466 Trace FindReachableObject<NLog.Layouts.Layout>:
+2023-12-27 19:41:41.0466 Trace Scanning FileTarget 'FileTarget(Name=logfile)'
+2023-12-27 19:41:41.0466 Trace  Scanning Property FileName '.\logs\${shortdate}_logfile.txt' NLog.Layouts.Layout
+2023-12-27 19:41:41.0466 Trace  Scanning Property Layout '${longdate} ${level:uppercase=true} ${message}' NLog.Layouts.Layout
+2023-12-27 19:41:41.0466 Trace FileTarget(Name=logfile) has 2 layouts
+2023-12-27 19:41:41.0466 Trace FileTarget(Name=logfile): Using appenderFactory: NLog.Internal.FileAppenders.SingleProcessFileAppender+Factory
+2023-12-27 19:41:41.0466 Info Configuration initialized.
