+++ Repository/RepositoryBase.cs
@@ -1,12 +1,33 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using Microsoft.EntityFrameworkCore;
+using System.Linq.Expressions;
 
 namespace Repository
 {
-    internal class IRepositoryBase
+    public class RepositoryBase<T> where T : class
     {
+        protected RepositoryContext RepositoryContext;
+
+        public RepositoryBase(RepositoryContext repositoryContext) =>
+            RepositoryContext = repositoryContext;
+
+        public IQueryable<T> FindAll(bool trackChanges) =>
+            !trackChanges ?
+                RepositoryContext.Set<T>()
+                .AsNoTracking() :
+                RepositoryContext.Set<T>();
+
+        public IQueryable<T> FindByCondition(Expression<Func<T, bool>> expression, bool trackChanges) =>
+            !trackChanges ?
+                RepositoryContext.Set<T>()
+                .Where(expression)
+                .AsNoTracking() :
+                RepositoryContext.Set<T>()
+                .Where(expression);
+
+        public void Create(T entity) => RepositoryContext.Set<T>().Add(entity);
+
+        public void Update(T entity) => RepositoryContext.Set<T>().Update(entity);
+
+        public void Delete(T entity) => RepositoryContext.Set<T>().Remove(entity);
     }
 }
